Ajusted return of ft_strlen, almost completed ft_split

+++ ft_strlen.c
@@ -6,14 +6,14 @@
 /*   By: romachad <romachad@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/21 06:59:21 by romachad          #+#    #+#             */
-/*   Updated: 2022/05/28 23:40:15 by romachad         ###   ########.fr       */
+/*   Updated: 2022/06/15 06:24:16 by romachad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != 0)
